• Sections 4.3 and 4.4 are implemented and marked complete in matrix-stats/req/v2.4.0-idiomaticGroovy.md:419. Public result scalars now use BigDecimal across the ANOVA,   contingency, and timeseries result types, and the public vector/matrix outputs that were still primitive are now Groovy-facing lists in places like matrix-stats/src/main/   groovy/se/alipsa/matrix/stats/contingency/Fisher.groovy, matrix-stats/src/main/groovy/se/alipsa/matrix/stats/timeseries/Ccm.groovy, matrix-stats/src/main/groovy/se/alipsa/   matrix/stats/timeseries/Johansen.groovy, and matrix-stats/src/main/groovy/se/alipsa/matrix/stats/timeseries/Df.groovy. I also removed the leftover @CompileStatic usage/   import from matrix-stats/src/main/groovy/se/alipsa/matrix/stats/formula/FormulaSupport.groovy per your correction, and fixed the redundant cast you pointed out in matrix-   stats/src/main/groovy/se/alipsa/matrix/stats/cluster/ClusteredPoint.groovy:132.
  The affected tests were updated to use list semantics instead of array assumptions in matrix-stats/src/test/groovy/contingency/FisherTest.groovy, matrix-stats/src/test/
  groovy/timeseries/CcmTest.groovy, and matrix-stats/src/test/groovy/timeseries/JohansenTest.groovy. Verification passed with ./gradlew :matrix-stats:test --tests 'AnovaTest'
  --tests 'contingency.*' --tests 'timeseries.*' :matrix-stats:codenarcMain :matrix-stats:codenarcTest (342 tests passed). One non-blocking warning remains from the custom
  switch-style check about 4 old case X: switches, but CodeNarc itself passed.

+++ matrix-stats/src/main/groovy/se/alipsa/matrix/stats/Anova.groovy
@@ -38,14 +38,14 @@ class Anova {
       categoryData << values
     }
 
-    result.fValue = FDistribution.oneWayAnovaFValue(categoryData) as Double
-    result.pValue = FDistribution.oneWayAnovaPValue(categoryData) as Double
+    result.fValue = FDistribution.oneWayAnovaFValue(categoryData)
+    result.pValue = FDistribution.oneWayAnovaPValue(categoryData)
     return result
   }
 
   static class AnovaResult {
-    Double pValue
-    Double fValue
+    BigDecimal pValue
+    BigDecimal fValue
 
     String toString() {
       return "pValue: ${pValue}, fValue: ${fValue}"
